+++ apache/httpd/srclib/apr-util/xml/apr_xml.c
@@ -1,9 +1,9 @@
-/* Copyright 2000-2005 The Apache Software Foundation or its licensors, as
- * applicable.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+/* Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -383,7 +383,7 @@ static apr_status_t do_parse(apr_xml_parser *parser,
         parser->error = APR_XML_ERROR_PARSE_DONE;
     }
     else {
-        int rv = XML_Parse(parser->xp, data, len, is_final);
+        int rv = XML_Parse(parser->xp, data, (int)len, is_final);
 
         if (rv == 0) {
             parser->error = APR_XML_ERROR_EXPAT;
